+++ dev-tools/omdb/tests/successes.out
@@ -5,7 +5,7 @@ stdout:
 Target Blueprint ......<REDACTED_BLUEPRINT_ID>....... @ nexus_generation: 1
 ID                                   |LAST_DRAINED_BLUEPRINT |STATE  |TRANSITIONING_TO 
 -------------------------------------+-----------------------+-------+-----------------
-..........<REDACTED_UUID>........... |n/a                    |active |                 
+..........<REDACTED_UUID>........... |                       |active |                 
 ---------------------------------------------
 stderr:
 note: using database URL postgresql://root@[::1]:REDACTED_PORT/omicron?sslmode=disable
